Intelligent name capitalization for the author mapper
